@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:mi_utem/models/asignatura.dart';
 
 class Horario {
-  List<Asignatura> asignaturas;
-  List<List<BloqueHorario>> horario;
-  List<dynamic> dias;
-  List<Periodo> periodos;
+  List<Asignatura>? asignaturas;
+  List<List<BloqueHorario>>? horario;
+  List<dynamic>? dias;
+  List<Periodo>? periodos;
 
   Horario({
     this.asignaturas,
@@ -14,7 +14,7 @@ class Horario {
     this.periodos
   });
 
-  factory Horario.fromJson(Map<String, dynamic> json) {
+  factory Horario.fromJson(Map<String, dynamic>? json) {
     if (json == null) {
       return Horario();
     }
@@ -37,16 +37,16 @@ class Horario {
     return list;
   }
 
-  List<String> get horasInicio {
-    return periodos.map((periodo) => periodo.horaInicio).toList();
+  List<String?> get horasInicio {
+    return periodos!.map((periodo) => periodo.horaInicio).toList();
   }
 
-  List<String> get horasIntermedio {
-    return periodos.map((periodo) => periodo.horaIntermedio).toList();
+  List<String?> get horasIntermedio {
+    return periodos!.map((periodo) => periodo.horaIntermedio).toList();
   }
 
-  List<String> get horasTermino {
-    return periodos.map((periodo) => periodo.horaTermino).toList();
+  List<String?> get horasTermino {
+    return periodos!.map((periodo) => periodo.horaTermino).toList();
   }
 
   List<List<BloqueHorario>> get horarioEnlazado {
@@ -54,15 +54,15 @@ class Horario {
 
     List<Color> colores = Colors.primaries.toList()..shuffle();
 
-    for (num i = 0; i < asignaturas.length; i++) {
-      var asignatura = asignaturas[i];
+    for (num i = 0; i < asignaturas!.length; i++) {
+      var asignatura = asignaturas![i as int];
       asignatura.colorAsignatura = colores[i];
     }
 
-    for (List<BloqueHorario> fila in horario) {
+    for (List<BloqueHorario> fila in horario!) {
       List<BloqueHorario> filaNueva = [];
       for (BloqueHorario bloque in fila) {
-        for (Asignatura asignatura in asignaturas) {
+        for (Asignatura asignatura in asignaturas!) {
           if (bloque.codigo == asignatura.codigo) {
             bloque.asignatura = asignatura;
           }
@@ -76,10 +76,10 @@ class Horario {
 }
 
 class Periodo {
-  String numero;
-  String horaInicio;
-  String horaIntermedio;
-  String horaTermino;
+  String? numero;
+  String? horaInicio;
+  String? horaIntermedio;
+  String? horaTermino;
 
   Periodo({
     this.numero,
@@ -88,7 +88,7 @@ class Periodo {
     this.horaTermino
   });
 
-  factory Periodo.fromJson(Map<String, dynamic> json) {
+  factory Periodo.fromJson(Map<String, dynamic>? json) {
     if (json == null) {
       return Periodo();
     }
@@ -113,9 +113,9 @@ class Periodo {
 }
 
 class BloqueHorario {
-  Asignatura asignatura;
-  String sala;
-  String codigo;
+  Asignatura? asignatura;
+  String? sala;
+  String? codigo;
 
   BloqueHorario({
     this.asignatura,
@@ -123,7 +123,7 @@ class BloqueHorario {
     this.codigo,
   });
 
-  factory BloqueHorario.fromJson(Map<String, dynamic> json) {
+  factory BloqueHorario.fromJson(Map<String, dynamic>? json) {
     if (json == null) {
       return BloqueHorario();
     }
@@ -134,7 +134,7 @@ class BloqueHorario {
     );
   }
 
-  static List<List<BloqueHorario>> fromJsonMatrix(dynamic json) {
+  static List<List<BloqueHorario>>? fromJsonMatrix(dynamic json) {
     if (json == null) {
       return null;
     }
