@@ -16,18 +16,19 @@ class AsignaturaScreen extends StatefulWidget {
   final Asignatura asignatura;
 
   AsignaturaScreen({
-    Key key,
-    this.asignatura,
+    Key? key,
+    required this.asignatura,
   }) : super(key: key);
 
   @override
   _AsignaturaScreenState createState() => _AsignaturaScreenState();
 }
 
-class _AsignaturaScreenState extends State<AsignaturaScreen> with SingleTickerProviderStateMixin {
-  TabController _tabController;
+class _AsignaturaScreenState extends State<AsignaturaScreen>
+    with SingleTickerProviderStateMixin {
+  TabController? _tabController;
   int _selectedIndex = 1;
-  Asignatura _asignaturaConNotas;
+  Asignatura? _asignaturaConNotas;
   List<Widget> _tabs = [];
 
   @override
@@ -36,13 +37,15 @@ class _AsignaturaScreenState extends State<AsignaturaScreen> with SingleTickerPr
     ReviewService.addScreen("AsignaturaScreen");
     _tabs = [
       AsignaturaResumenTab(asignatura: widget.asignatura),
-      AsignaturaNotasTab(asignatura: widget.asignatura, onNotas: _onAsignaturaConNotas),
+      AsignaturaNotasTab(
+          asignatura: widget.asignatura, onNotas: _onAsignaturaConNotas),
       AsignaturaEstudiantesTab(asignatura: widget.asignatura),
     ];
-    _tabController = TabController(length: _tabs.length, vsync: this, initialIndex: 1);
-    _tabController.addListener(() {
+    _tabController =
+        TabController(length: _tabs.length, vsync: this, initialIndex: 1);
+    _tabController!.addListener(() {
       setState(() {
-        _selectedIndex = _tabController.index;
+        _selectedIndex = _tabController!.index;
       });
     });
   }
@@ -60,34 +63,36 @@ class _AsignaturaScreenState extends State<AsignaturaScreen> with SingleTickerPr
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        appBar: CustomAppBar(
-          title:
-              Text("${widget.asignatura.nombre}"),
-          
-          actions: _mostrarCalculadora ? [
-            IconButton(
-              icon: Icon(Mdi.calculator),
-              tooltip: "Calculadora",
-              onPressed: () {
-                Get.to(
-                  CalculadoraNotasScreen(asignaturaInicial: _asignaturaConNotas),);
-              },
-            ),
-          ] : [],
-          bottom: TabBar(
-            controller: _tabController,
-            indicatorColor: Colors.white.withOpacity(0.8),
-            tabs: [
-              Tab(text: "Resumen"),
-              Tab(text: "Notas"),
-              Tab(text: "Estudiantes"),
-            ].sublist(0, _tabs.length),
-          ),
-        ),
-        body: TabBarView(
+      appBar: CustomAppBar(
+        title: Text("${widget.asignatura.nombre}"),
+        actions: _mostrarCalculadora
+            ? [
+                IconButton(
+                  icon: Icon(Mdi.calculator),
+                  tooltip: "Calculadora",
+                  onPressed: () {
+                    Get.to(
+                      CalculadoraNotasScreen(
+                          asignaturaInicial: _asignaturaConNotas),
+                    );
+                  },
+                ),
+              ]
+            : [],
+        bottom: TabBar(
           controller: _tabController,
-          children: _tabs,
+          indicatorColor: Colors.white.withOpacity(0.8),
+          tabs: [
+            Tab(text: "Resumen"),
+            Tab(text: "Notas"),
+            Tab(text: "Estudiantes"),
+          ].sublist(0, _tabs.length),
         ),
+      ),
+      body: TabBarView(
+        controller: _tabController,
+        children: _tabs,
+      ),
     );
   }
 }
