@@ -4,7 +4,6 @@ import 'dart:io';
 import 'package:awesome_notifications/awesome_notifications.dart';
 import 'package:firebase_analytics/firebase_analytics.dart';
 import 'package:firebase_remote_config/firebase_remote_config.dart';
-import 'package:flushbar/flushbar.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
@@ -28,7 +27,7 @@ import 'package:mi_utem/widgets/custom_app_bar.dart';
 
 class MainScreen extends StatefulWidget {
   final Usuario usuario;
-  MainScreen({Key key, @required this.usuario}) : super(key: key);
+  MainScreen({Key? key, required this.usuario}) : super(key: key);
 
   @override
   _MainScreenState createState() => _MainScreenState();
@@ -37,7 +36,7 @@ class MainScreen extends StatefulWidget {
 class _MainScreenState extends State<MainScreen> {
   int _a = 0;
 
-  RemoteConfig _remoteConfig;
+  RemoteConfig? _remoteConfig;
 
   @override
   void initState() {
@@ -58,7 +57,7 @@ class _MainScreenState extends State<MainScreen> {
 
     ReviewService.addScreen("MainScreen");
     ReviewService.checkAndRequestReview();
-    SchedulerBinding.instance.addPostFrameCallback((_) {
+    SchedulerBinding.instance!.addPostFrameCallback((_) {
       _checkAndPerformUpdate(context);
     });
   }
@@ -149,7 +148,7 @@ class _MainScreenState extends State<MainScreen> {
                                   overflow: TextOverflow.ellipsis,
                                   style: Theme.of(context)
                                       .textTheme
-                                      .headline3
+                                      .headline3!
                                       .copyWith(
                                         color: Colors.white,
                                       ),
@@ -192,7 +191,7 @@ class _MainScreenState extends State<MainScreen> {
                                   overflow: TextOverflow.ellipsis,
                                   style: Theme.of(context)
                                       .textTheme
-                                      .headline3
+                                      .headline3!
                                       .copyWith(
                                         color: Colors.white,
                                       ),
@@ -223,11 +222,11 @@ class _MainScreenState extends State<MainScreen> {
                 });
                 if (_a >= 11) {
                   _a = 0;
-                  if (_remoteConfig.getBool(ConfigService.EG_HABILITADOS)) {
-                    Flushbar(
-                      message: _remoteConfig.getString(ConfigService.PRONTO_EG),
+                  if (_remoteConfig!.getBool(ConfigService.EG_HABILITADOS)) {
+                    /* Flushbar(
+                      message: _remoteConfig!.getString(ConfigService.PRONTO_EG),
                       duration: Duration(seconds: 3),
-                    )..show(context);
+                    )..show(context); */
                     FirebaseAnalytics().logEvent(name: "pronto_eg");
                   }
                 }
@@ -240,7 +239,8 @@ class _MainScreenState extends State<MainScreen> {
                   children: <Widget>[
                     Icon(
                       IconData(
-                          _remoteConfig.getInt(ConfigService.HOME_PRONTO_ICONO),
+                          _remoteConfig!
+                              .getInt(ConfigService.HOME_PRONTO_ICONO),
                           fontFamily: 'MaterialIcons'),
                       size: 150,
                       color: Colors.grey,
@@ -249,7 +249,7 @@ class _MainScreenState extends State<MainScreen> {
                       height: 20,
                     ),
                     Text(
-                      _remoteConfig.getString(ConfigService.HOME_PRONTO_TEXTO),
+                      _remoteConfig!.getString(ConfigService.HOME_PRONTO_TEXTO),
                       style: TextStyle(color: Colors.grey, fontSize: 16),
                     )
                   ],
