@@ -254,13 +254,13 @@ class ConfigService {
     ),
   };
 
-  static ConfigService _instance;
+  static ConfigService? _instance;
 
-  static RemoteConfig _remoteConfig;
+  static RemoteConfig? _remoteConfig;
 
-  static RemoteConfig get config => _remoteConfig;
+  static RemoteConfig? get config => _remoteConfig;
 
-  static Future<ConfigService> getInstance() async {
+  static Future<ConfigService?> getInstance() async {
     if (_instance == null) {
       _remoteConfig = RemoteConfig.instance;
       await initialise();
@@ -272,8 +272,8 @@ class ConfigService {
 
   static Future initialise() async {
     try {
-      await _remoteConfig.setDefaults(defaults);
-      await _remoteConfig.fetchAndActivate();
+      await _remoteConfig!.setDefaults(defaults);
+      await _remoteConfig!.fetchAndActivate();
     } catch (exception) {
       print(
           'Unable to fetch remote config. Cached or default values will be used');
@@ -282,11 +282,11 @@ class ConfigService {
 
   static Future update() async {
     try {
-      await _remoteConfig.setConfigSettings(RemoteConfigSettings(
+      await _remoteConfig!.setConfigSettings(RemoteConfigSettings(
           minimumFetchInterval: Duration.zero,
           fetchTimeout: Duration(minutes: 1)));
-      await _remoteConfig.fetchAndActivate();
-      await _remoteConfig.setConfigSettings(RemoteConfigSettings(
+      await _remoteConfig!.fetchAndActivate();
+      await _remoteConfig!.setConfigSettings(RemoteConfigSettings(
           minimumFetchInterval: Duration(hours: 12),
           fetchTimeout: Duration(minutes: 1)));
     } catch (exception) {
