@@ -9,7 +9,7 @@ extension StringExtension on Color {
 }
 
 class AsistenciaChart extends StatelessWidget {
-  final Asistencia asistencia;
+  final Asistencia? asistencia;
 
   AsistenciaChart({this.asistencia});
 
@@ -17,9 +17,9 @@ class AsistenciaChart extends StatelessWidget {
   List<charts.Series<GaugeSegment, String>> _procesarDatos() {
     
     final data = [
-      new GaugeSegment('Asistido', asistencia.asistidos, charts.Color.fromHex(code: MainTheme.aprobadoColor.toHex())),
+      new GaugeSegment('Asistido', asistencia!.asistidos as int, charts.Color.fromHex(code: MainTheme.aprobadoColor.toHex())),
       //new GaugeSegment('No asistido', asistencia.noAsistidos, charts.Color.fromHex(code: MainTheme.reprobadoColor.toHex())),
-      new GaugeSegment('Sin registro', asistencia.sinRegistro, charts.Color.fromHex(code: MainTheme.disabledColor.toHex())),
+      new GaugeSegment('Sin registro', asistencia!.sinRegistro as int, charts.Color.fromHex(code: MainTheme.disabledColor!.toHex())),
     ];
 
     return [
@@ -35,8 +35,8 @@ class AsistenciaChart extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    print(asistencia.asistidos);
-    print(asistencia.sinRegistro);
+    print(asistencia!.asistidos);
+    print(asistencia!.sinRegistro);
     return Column(
       crossAxisAlignment: CrossAxisAlignment.center,
       children: [
@@ -73,7 +73,7 @@ class AsistenciaChart extends StatelessWidget {
               width: 15
             ),
             Container(width: 10),
-            Text("Asistidos (${asistencia.asistidos})"),
+            Text("Asistidos (${asistencia!.asistidos})"),
           ],
         ),
         Container(height: 5),
@@ -90,7 +90,7 @@ class AsistenciaChart extends StatelessWidget {
               width: 15
             ),
             Container(width: 10),
-            Text("No asistidos (${asistencia.noAsistidos})"),
+            Text("No asistidos (${asistencia!.noAsistidos})"),
           ],
         ),
         Container(height: 5),
@@ -107,7 +107,7 @@ class AsistenciaChart extends StatelessWidget {
               width: 15
             ),
             Container(width: 10),
-            Text("Sin registro (${asistencia.sinRegistro})"),
+            Text("Sin registro (${asistencia!.sinRegistro})"),
           ],
         ),
       ],
