@@ -25,9 +25,9 @@ class HorarioService {
 
       Horario horario = Horario.fromJson(response.data);
 
-      for (var dia in horario.horario) {
+      for (var dia in horario.horario!) {
         for (var bloque in dia) {
-          print("HorarioService bloque ${bloque?.codigo}");
+          print("HorarioService bloque ${bloque.codigo}");
         }
       }
 
